I made the pages work independently
It is small and I haven't made much progress in the past 5 hours but I realized I had 2 click event listeners for all of the pages which was making everything a lot more complicated. After reviewing the code thoroughly, I finally fixed everything and I am back on my path to making the restaurant page have tabbed pages. Its time to focus on the bigger picture!