tabbed button solution
I finally came to what appears to be a solid solution to making my buttons tabbed for the restaurant page's overall functionality. After 6-8 hours of figuring things out I finally realized what I've been doing wrong this entire time. I've been trying to get the separate divs within the other files to work as I had them imported but I didn't realize the function had to be called as well for the divs to work. I called the function but the content cluttered the main page so within the remove items function I made the display go away by default and only show when the corresponding button is clicked.